Move SetFromList as static inner class of BinaryJedis
